requirements.txt
fastapi
uvicorn
azure-ai-contentsafety
azure-ai-textanalytics
openai
python-dotenv